agents/content_builder/pyproject.toml
[project]
name = "content-builder-agent"
version = "0.1.0"
description = "Content Builder microservice agent"
requires-python = ">=3.13"
dependencies = [
    "shared-utils @ file:///${PROJECT_ROOT}/shared",
    "google-adk",
    "a2a-sdk",
    "fastapi",
    "uvicorn",
    "python-json-logger",
    "google-cloud-logging",
    "opentelemetry-exporter-gcp-trace",
]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
